database/requirements.txt
h5py==3.1.0
psycopg2-binary==2.8.3
tqdm==4.31.1
